refactor: move `test_subscriber` into `flox_rust_sdk`
# Conflicts:
#	cli/flox-rust-sdk/src/providers/catalog.rs
