+++ cli/flox/Cargo.toml
@@ -62,6 +62,7 @@ serial_test.workspace = true
 temp-env.workspace = true
 tempfile.workspace = true
 toml.workspace = true
+flox-rust-sdk = { workspace = true, features = ["tests"] }
 
 
 [features]
