+++ cli/flox/src/commands/activate.rs
@@ -29,7 +29,7 @@ use flox_rust_sdk::models::environment::{
 use flox_rust_sdk::providers::build::FLOX_RUNTIME_DIR_VAR;
 use flox_rust_sdk::providers::services::shutdown_process_compose_if_all_processes_stopped;
 use flox_rust_sdk::providers::upgrade_checks::UpgradeInformationGuard;
-use flox_rust_sdk::utils::traceable_path;
+use flox_rust_sdk::utils::logging::traceable_path;
 use indoc::{formatdoc, indoc};
 use itertools::Itertools;
 use log::{debug, warn};
@@ -880,15 +880,13 @@ mod tests {
 
 #[cfg(test)]
 mod upgrade_notification_tests {
-    use std::io::Write;
-    use std::sync::{Arc, Mutex};
-
     use flox_rust_sdk::flox::test_helpers::flox_instance;
     use flox_rust_sdk::models::environment::path_environment::test_helpers::new_path_environment_from_env_files;
     use flox_rust_sdk::models::environment::UpgradeResult;
     use flox_rust_sdk::models::lockfile::LockedPackage;
     use flox_rust_sdk::providers::catalog::GENERATED_DATA;
     use flox_rust_sdk::providers::upgrade_checks::UpgradeInformation;
+    use flox_rust_sdk::utils::logging::test_helpers::CollectingWriter;
     use time::OffsetDateTime;
     use tracing::Subscriber;
     use tracing_subscriber::filter::FilterFn;
@@ -896,45 +894,11 @@ mod upgrade_notification_tests {
 
     use super::*;
 
-    #[derive(Debug, Default)]
-    struct CollectingWriter {
-        buffer: Mutex<Vec<u8>>,
-    }
-
-    impl Display for CollectingWriter {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            let buffer = self.buffer.lock().unwrap();
-            let str_content = String::from_utf8_lossy(&buffer);
-            write!(f, "{str_content}")
-        }
-    }
-    impl Write for &CollectingWriter {
-        fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-            self.buffer.lock().unwrap().write(buf)
-        }
-
-        fn flush(&mut self) -> std::io::Result<()> {
-            self.buffer.lock().unwrap().flush()
-        }
-    }
-
-    // For now this is a POC of using tracing for output tests,
-    // evenatually we should probably move that to the tracing utils or `message` module.
-    fn test_subscriber() -> (impl Subscriber, Arc<CollectingWriter>) {
-        let writer = Arc::new(CollectingWriter::default());
-
-        // TODO: also tee to test output?
-        let subscriber = tracing_subscriber::fmt()
-            .with_writer(writer.clone())
-            .compact()
-            .without_time()
-            .with_level(false)
-            .with_target(false)
-            .finish()
-            .with(FilterFn::new(|metadata| {
-                metadata.target() == "flox::utils::message"
-            }));
-
+    fn test_subscriber() -> (impl Subscriber, CollectingWriter) {
+        let (subscriber, writer) = flox_rust_sdk::utils::logging::test_helpers::test_subscriber();
+        let subscriber = subscriber.with(FilterFn::new(|metadata| {
+            metadata.target() == "flox::utils::message"
+        }));
         (subscriber, writer)
     }
